requirements_test.txt
coverage==4.4.1
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0
sphinx_rtd_theme
sphinx
recommonmark

django-model-utils>=2.0

# Additional test requirements go here
